requirements.txt
imagecodecs==2020.2.18
matplotlib==3.2.1
imageio==2.8.0
opencv-python==4.2.0.32
pandas==1.0.3
Pillow==6.1.0
pydicom==1.3.0
torch==1.2.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.4.0
tqdm==4.32.1
absl-py==0.7.1
